create drag and drop component
